refactor: Add UrlReplacer class for testability
